pyproject.toml
[tool.poetry]
name = "griptape-cli"
version = "0.1.0"
description = "CLI for the Griptape Framework and Cloud"
authors = ["Griptape <hello@griptape.ai>"]
license = "Apache 2.0"
readme = "README.md"
repository = "https://github.com/griptape-ai/griptape-cli"

packages = [
    {include = "griptape"}
]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
griptape = { git = "https://github.com/griptape-ai/griptape.git", branch = "main" }
cookiecutter = "^2.1.1"
attrs = "^23.1.0"
stringcase = "^1.2.0"

[tool.poetry.group.test.dependencies]
pytest = "*"
pytest-mock = "*"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gt = "griptape.cli.main:cli"
